# tests/_data/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Elexp Conclusions API",
    "version": "1.0"
  },
  "paths": {
...
    "/conclusion/{conclusion_id}/paragraph/{paragraph_id}/block": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "ID заключения",
            "format": "uuid",
            "in": "path",
            "name": "conclusion_id",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID раздела",
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateConclusionBlockDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/ApiResponse"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/ConclusionBlock"
                    }
                  },
                  "type": "object"
                }
              ]
            }
          }
        },
        "tags": [
          "Conclusion Block"
        ],
        "operationId": "App\\Http\\Controller\\Conclusion\\BlockController::createBlock",
        "summary": "Создать блок"
      }
    },
    "/conclusion/{conclusion_id}/paragraph/{paragraph_id}/block/{block_id}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "ID заключения",
            "format": "uuid",
            "in": "path",
            "name": "conclusion_id",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID раздела",
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "ID блока",
            "in": "path",
            "name": "block_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        },
        "tags": [
          "Conclusion Block"
        ],
        "operationId": "App\\Http\\Controller\\Conclusion\\BlockController::deleteBlock",
        "summary": "Удалить блок"
      }
    },
...
    "/conclusion/{conclusion_id}/paragraph/{paragraph_id}/block/{block_id}/executor": {
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "ID заключения",
            "format": "uuid",
            "in": "path",
            "name": "conclusion_id",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID раздела",
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "ID блока",
            "in": "path",
            "name": "block_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SetBlockExecutorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/ApiResponse"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/ConclusionBlock"
                    }
                  },
                  "type": "object"
                }
              ]
            }
          }
        },
        "tags": [
          "Conclusion Block"
        ],
        "operationId": "App\\Http\\Controller\\Conclusion\\BlockController::setExecutor",
        "summary": "Изменить исполнителя блока"
      }
    },
    "/conclusion/{conclusion_id}/paragraph/{paragraph_id}/block/{block_id}/kind": {
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "ID заключения",
            "format": "uuid",
            "in": "path",
            "name": "conclusion_id",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID раздела",
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "ID блока",
            "in": "path",
            "name": "block_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ConclusionBlockSetKindDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/ApiResponse"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/ConclusionBlock"
                    }
                  },
                  "type": "object"
                }
              ]
            }
          }
        },
        "tags": [
          "Conclusion Block"
        ],
        "operationId": "App\\Http\\Controller\\Conclusion\\BlockController::setKind",
        "summary": "Изменить тип блока"
      }
    },
    "/conclusion/{conclusion_id}/paragraph/{paragraph_id}/block/{block_id}/state": {
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "ID заключения",
            "format": "uuid",
            "in": "path",
            "name": "conclusion_id",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID раздела",
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "ID блока",
            "in": "path",
            "name": "block_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChangeConclusionBlockStateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/ApiResponse"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/ConclusionBlock"
                    }
                  },
                  "type": "object"
                }
              ]
            }
          }
        },
        "tags": [
          "Conclusion Block"
        ],
        "operationId": "App\\Http\\Controller\\Conclusion\\BlockController::setState",
        "summary": "Изменить статус блока"
      }
    },
...
  },
  "definitions": {
...
    "ApiResponse": {
      "properties": {
        "data": {},
        "errors": {
          "items": {
            "type": "string"
          },
          "type": "array",
          "x-nullable": true
        },
        "state": {
          "description": "Ok or failed request",
          "enum": [
            "ok",
            "error"
          ],
          "type": "string"
        }
      },
      "type": "object"
    },
    "Certificate": {
      "properties": {
        "scope": {
          "nullable": false,
          "type": "string"
        }
      },
      "type": "object"
    },
    "ChangeConclusionBlockStateDTO": {
      "properties": {
        "new_state": {
          "$ref": "#/definitions/ConclusionBlockState"
        }
      },
      "type": "object"
    },
...
    "ConclusionBlock": {
      "properties": {
        "executor": {
          "$ref": "#/definitions/User"
        },
        "file_path": {
          "type": "string",
          "x-nullable": true
        },
        "id": {
          "type": "integer"
        },
        "kind": {
          "$ref": "#/definitions/ConclusionBlockKind"
        }
      },
      "type": "object"
    },
...
    "ConclusionBlockKind": {
      "enum": [
        "text",
        "dict"
      ],
      "type": "string"
    },
    "ConclusionBlockSetKindDTO": {
      "properties": {
        "kind": {
          "$ref": "#/definitions/ConclusionBlockKind"
        }
      },
      "type": "object"
    },
    "ConclusionBlockState": {
      "enum": [
        "waiting_to_start",
        "work_in_progress",
        "sent_to_review",
        "on_review",
        "declined",
        "completed"
      ],
      "type": "string"
    },
...
    "CreateConclusionBlockDTO": {
      "properties": {
        "kind": {
          "$ref": "#/definitions/ConclusionBlockKind"
        }
      },
      "type": "object"
    },
...
    "SetBlockExecutorDTO": {
      "properties": {
        "user_id": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
    "User": {
      "properties": {
        "certificates": {
          "items": {
            "$ref": "#/definitions/Certificate"
          },
          "type": "array"
        },
        "email": {
          "$ref": "#/definitions/UserEmail"
        },
        "full_name": {
          "description": "Full Name",
          "type": "string"
        },
        "id": {
          "nullable": false,
          "type": "integer"
        },
        "role": {
          "$ref": "#/definitions/UserRole"
        }
      },
      "type": "object"
    },
...
    "UserEmail": {
      "format": "email",
      "type": "string"
    },
    "UserRole": {
      "enum": [
        "client",
        "verifier",
        "admin",
        "expert",
        "project_manager"
      ],
      "type": "string"
    }
  },
...
}